requirements_clean.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
boto3==1.38.32
botocore==1.38.32
Brotli==1.1.0
click==8.2.1
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
distro==1.9.0
doclayout_yolo==0.0.2b1
fast-langdetect==0.2.5
fasttext-predict==0.9.2.4
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.58.2
frozenlist==1.6.2
fsspec==2024.6.1
ftfy==6.3.1
hf-xet==1.1.3
huggingface-hub==0.32.4
humanfriendly==10.0
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.8
loguru==0.7.3
magic-pdf==1.3.12
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
onnxruntime==1.22.0
openai==1.84.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
pandas==2.3.0
peft==0.15.2
pillow==11.0.0
propcache==0.3.1
protobuf==6.31.1
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.24.14
pyparsing==3.2.3
PyQt6==6.7.1
python-dotenv==1.1.0
pytz==2025.2
rapid-table==1.0.5
regex==2024.11.6
robust-downloader==0.0.2
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
setuptools==78.1.1
shapely==2.1.1
simsimd==6.4.9
stringzilla==3.12.5
sympy==1.13.3
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers==4.52.4
triton==3.3.1
trl==0.18.1
tzdata==2025.2
ultralytics==8.3.152
ultralytics-thop==2.0.14
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.0
